pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "unifi-lab-kit"
version = "0.1.0"
description = "Reproducible reset-and-reconfigure toolkit for a UniFi USG + Controller + Ubuntu-servers + NAS lab network."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "unifi-lab-kit contributors" },
]
keywords = ["unifi", "usg", "netdevops", "network-automation", "ssh", "samba", "homelab"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Development Status :: 4 - Beta",
]
dependencies = [
    "paramiko>=3.4",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.32",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
"Source" = "https://github.com/thc1006/unifi-lab-kit"
"Issues" = "https://github.com/thc1006/unifi-lab-kit/issues"

[project.scripts]
ulk-usg = "unifi_lab_kit.usg:main"
ulk-controller = "unifi_lab_kit.controller:main"
ulk-scan = "unifi_lab_kit.scan:main"
ulk-deploy-key = "unifi_lab_kit.ssh_deploy:main"
ulk-static-ip = "unifi_lab_kit.static_ip:main"
ulk-samba = "unifi_lab_kit.nas_samba:main"
ulk-verify = "unifi_lab_kit.verify:main"

[tool.hatch.build.targets.wheel]
packages = ["src/unifi_lab_kit"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]
# archive/ is frozen history; _secrets/ is local-only and gitignored.
# Keeping both out of lint so CI doesn't fight legacy code or private backups.
extend-exclude = ["archive", "_secrets"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]  # line length

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
files = ["src"]
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
# Lenient for initial release — paramiko's stubs are weak and much of
# the codebase talks to remote shells whose return types are str-only
# by construction. Tighten later module-by-module.
check_untyped_defs = false
disallow_untyped_defs = false
